requirements.txt
azure-iot-device==2.8.0
can==0.0.0
certifi==2021.10.8
charset-normalizer==2.0.7
deprecation==2.1.0
idna==3.3
janus==0.4.0
packaging==21.2
paho-mqtt==1.6.1
psutil==5.8.0
pyparsing==2.4.7
PySocks==1.7.1
requests==2.26.0
requests-unixsocket==0.2.0
six==1.16.0
urllib3==1.26.7
